frontend: restore GraphQL playground
I thought we could get away with using django-graphene's GraphiQL
integration, but that has security issues. Using the React GraphQL
playground integrates seamlessly with our authentication, so import the
code for that from the old repository. Also fix the layout issues, so
that the playground is not rendered inside the same chrome as the rest
of the application is.