Avoid hardcoding Cognito configuration
Plumb Pulumi's Cognito outputs through both the backend and the
frontend. This is tricky in development, since Pulumi doesn't manage
local dev environments; instead, provide a script that hardcodes the
outputs of the staging stack in the repository. This isn't ideal, but
it's the best I have to offer at the moment.

This work exposed that the frontend "public" directory (containing an
index.html template and some favicons) was not getting served. This
commit fixes that by making Django responsible for serving all static
files except for the JavaScript bundle.