Fix caching behavior of frontend NPM packages
We discussed the issue on Slack a while back. The tl;dr is that
the NPM package caching was broken with Docker Compose, so running
`docker-compose up --build` after adding a new NPM package did not work
until you blew away the existing fe-node-modules volume.

This commit reworks how we build NPM packages into the frontend image to
avoid this hazard. The approach is described in detail within the patch.