Build & ship sourcemaps in the frontend dev server
This enables building sourcemaps in typescript's config, and enables
the fastest-building/rebuilding source map integration in the webpack
dev server.

The production source code should be unaffected by this, all we want
here is to get the sources such that stepping/debugging the TS
code (and CSS!) works in a reasonable way.