console: introduce Django REST Framework-backed API
This commit introduces Django REST Framework, and uses it to power our
telemetry API and a new HTTP-based "IsEmailAuthenticated" API. The goal
is to build this out into a full replacement for our GraphQL API, then
delete the GraphQL API. This commit is the first step in that process.

Extracted from #723.
Touches #640.