frontend: rewrite Cognito integration
Our AWS Cognito integration on the frontend was rather messy. It
depended on several different AWS Cognito libraries, only one of which
was actually used, which added hundreds of NPM packages we didn't need.
It involved two separate providers: one for authenticating with
Cognito, and one for providing those authentication credentials to our
GraphQL library. And the Cognito provider needlessly duplicated the
auth interface in the interest of abstraction, but the function calls
provided were so deeply Cognito specific that it wasn't a particularly
useful abstraction.

This commit rewrites the integration to be much simpler. There is now
just one `AuthProvider` which maintains the authentication state machine
(logged out, logged in with a user, or logged out with a possible error
message.) Components which need to perform more complicated Cognito API
calls now do so by making those calls against the Cognito API directly.

The `AuthProvider` now handles integration credentials to our GraphQL
library, rather than offloading that to a *separate* provider, and will
be an easy spot to integrate credential handling for the forthcoming
REST API.

This commit also removes the tests of the AuthProvider because they were
unnecessarily brittle and didn't test anything interesting. (The login
test, for example, just tested that the <Login> component rendered a
username and password form field, but we already test that logging in
*actually works*, Cognito and all, via the integration tests.)