config: centralize management of Materialize releases
Rework how we track the available Materialize releases so that the
`config.mz_releases` module is in charge. This module now hosts a
`RELEASES` registry that maps Materialize versions to their Docker
digests, which is the API that the rest of the codebase wants. It also
uses the recently-introduced pydantic library to handle the parsing of
the JSON file based on the MzRelease dataclass.

Various other improvements include:

  * A proper script for blessing a new Materialize release, rather than
    a bunch of complicated jq code that has to be copied and pasted out
    of a Markdown file.

  * Removal of the telemetry update loop, which pulled the latest
    release from the docs website. Since teaching Materialize Cloud
    about new releases is now part of the official release process, the
    telemetry server can just report our hardcoded knowledge of the
    latest release. (This information just powers the new version
    notification in materialized, so there's no harm if it's a bit
    stale.)

This commit was extracted from #750 so we could consider these code
improvements independently from the question of whether to use Docker
tags or digests in deployments.