rework of deployment state machine
The database now only stores the desired state, not the current state.

There are two new flags, used to trigger the reconciler:
* flagged_for_deletion
* flagged_for_update

When one of these flags is set in the database, the reconciler
(deployment_controller) will perform Kubernetes operations to reach the
desired state, then unset the flag.

There are now fewer states presented to the user:
States from the new flags:
* Destroying
* Updating

States from Kubernetes:
* Healthy
* Starting
* Damaged

Catch all:
* Unknown

The upgrade button is now shown whenever the version doesn't match,
regardless of the state of the cluster.

New user email and slack notifications are removed due to difficulty and
the fact they will be removed soon for the Frontegg change anyway.

Resolves #803