Tailscale support
Adds support for creating Tailscale containers in the same pod as
materialized, using an auth key supplied by the user.

The Tailscale auth key and state are stored in K8S secrets, and both are
removed on deletion or an update that removes Tailscale.

The pod will get an additional security group allowing ingress on
41641/udp (the Tailscale well-known NAT traversal port).
The cluster security group will allow 443/tcp from the Tailscale
security group.

Docker image and ECR repo creation are moved to a function, and we push
the Tailscale image to an ECR repo.

Kubernetes does not trigger recreation of a pod when its configmaps or
secrets change, so this PR also adds a sha256 sum of the config and
secret objects to the statefulset annotations, which does trigger
recreation of the pod on update.

Known issues/limitations:

* The `tailscale up` command is run with no supervision or init.

* The pod does not fail if Tailscale authentication fails.

* Liveness/Readiness probes do not yet exist.

* There is no API to get the Tailscale logs or for the user to otherwise
  debug it.

* Tailscale doesn't work with minikube, since it doesn't have anywhere
  to pull the image from.

* The enable_tailscale API is a bit clunky.