Wait until the frontend shows the new version before trying psql
The upgrade test would time out in psql connections because it tried
connecting too early: After hitting the upgrade button, it would often
see a "Healthy" status on the old version, and then attempt to dial
into the psql port (which then belonged to the old container's IP
address which was probably shut down... and then there's DNS update
delays).

So, first assert that the frontend is showing the right version before
assuming that a "healthy" status applies to a successful upgrade.