Add UI for user index disabling mode
"User Indexes Disabled" is now a status badge a deployment can have.
If the deployment is not otherwise OK, it'll still show the relevant
other badge instead.

The disableUserIndexes badge is blue and includes a helpful tooltip.

On the detail view, a banner above the connection instructions emphasizes
that the deployment is in disableUserIndexes mode. Then people know what to
expect when they connect.

The deployment update modal lets you turn disableUserIndexes mode on and off.
Includes a new Formik-wrapped format component, SwitchField.