Update restful-react library from 15.9.2 to 15.9.4
It appears that locally, Docker was not installing the new version of restful-react,
despite it being specified in package.json and package-lock.json. I had
to go manually into the frontend container and frob it. If we're using `npm ci`
instead of `npm install` on CI, that would give it a truly fresh start and
possibly explain the differing behavior.

I can't explain docker/npm's behavior though.