Add environments dropdown
This patch introduces a new state library, Recoil. This will let us
set the environment app-wide without drilling the environment prop down to
every page and piece of the site that will need it. React context can also
do this, but Recoil has some nice performance properties and a friendlier
API imo.

The environment dropdown's default option is 'All'. The unfiltered option
for now seemed less user-hostile than to filter the deployment list on
initial view. This is temporary, till we move it to the top bar and truly
restrict the views to specific environments.

The "environments" are currently a hacky combo of providers and regions.
Eventually they'll have their own backend endpoint we can query.