Move deployment logs modal to top of detail page
Marta's feedback was that she couldn't find it all the way in the footer
of the metrics card (which is also at the very bottom of the page, aggravating
the issue). Let's move it up!

I put it next to the Edit and Destroy buttons. It's not a modifying action,
but it is "admin-y", and I gave it secondary (outlined) styling.