Add list pages for platform nouns
Recoil tracks whether you're in platform or not (which determines
the available routes and the contents of the nav bar).

At smaller screen widths, the tabs in the nav collapse into a
hamburger menu. If we add more stuff to the nav bar (e.g. the enviro
dropdown) the break points here may need adjustment.

Currently the source, views, sinks, and clusters tables are all empty.