Use "sslmode=verify-full" instead of "require" for materialized conns
Turns out that sslmode=require checks only that there is SSL on the
wire. It does _not_ check that the certificate is valid, or that the
hostname you're connecting to is even certified by the
certificate. This allows easy MITM attacks.

Instead, let's tell users to use sslmode=verify-full (and use it for
ourselves). This one _does_ check that the cert is valid, and that the
hostname matches.