Don't try to version-compare non-semver deployment versions
It's possible for users (with dangerous features on) to manually spin
up mzcloud deployments that don't have semver versions (docker tags
like
materialize/materialized:mzbuild-OXSVH2SFPMVSKK4PHK4RWOLGCOYCHPMU).

We should not offer to upgrade those, and more than that - we
shouldn't even try to semver-compare them, as the version comparison
library that we use throws an error, and the page fails to render.