Use node port for deployments in local dev env
In local development environments, access deployments by their floating
node port, rather than mandating that they run on port 6875. This has
two major benefits:

  * You can actually run more than one deployment, because the
    deployment doesn't take up the one available port 6875 on the host.

  * We no longer need to run `minikube tunnel`; the services can be
    accessed directly via `MINIKUBE_HOST:SERVICE_PORT`. On macOS, you
    still need to forward node ports to the Minikube VM, but that's easy
    to do via a new `bin/port-forward` script.

The upshot is that local development on an M1 Mac is now as seamless as
local development on an Intel mac.

The new node port-forwarding strategy will be reused to facilitate local
development of the new platform APIs.