frontend: update platform tabs
We're punting on the clusters, sources, sinks, and views tabs for now.
Remove them from the nav bar. Also add the legacy deployments tab with a
"LEGACY" label, so that existing customers who want to trial the new
platform experience have a way to move between the two worlds.

Then, add a new "regions" tab. This is a temporary tab that will be used
to manage regions before we've got the dashboard tab built out. I'll
fill this tab in in a forthcoming commit.