frontend: wire up basic auto-refreshing clusters view
Add back the "clusters" tab in the platform UI, and wire it up so that
it auto-refreshes the list of clusters every 5s, like it did in last
week's demo. Again, this tab will need to undergo heavy refinement;
adding it now is mostly useful as a demonstration of how to run a SQL
query against an environment.

Touches #2470.