Create cluster in e2e tests
Adjust the end-to-end tests to actually create a cluster and run a query
in it. This will catch many of the bugs we've been seeing lately in
which the environment turns on but cluster creation does not work.

Also fix a bug in the end-to-end test that caused it to declare success
after 5m of trying to connect, regardless of whether it had successfully
connected or not. This bug is why several PRs that broke environment
deployment slipped through CI.

Finally, make the cluster replica size map vary based on Pulumi stack.
Personal, staging, and production stacks all have different instance
sizes, so they need different size maps, as do local development
environments in Minikube.