e2e-tests: smoke test S3 sources
Extend the platform end-to-end test to create an S3 source for a bucket
with known contents. This gives us test coverage on several fronts:

  * Outbound network access from storage pods
  * The basic function of S3 sources
  * AWS IAM role assumption

This test exposed that `materialized` was not creating the `storaged`
and `computed` pods with the correct service account; that's fixed in
this commit as well.

Fix #2915.