Light mode styling and other refinements on home page
We use the default card background rather than a hardcoded grey for the
background of the home page steps accordion.

The opened item(s) in the list are highlighted and there's a larger indent
for the contained content.

Connect steps have the default borders between them.
The borderless accordion in the edit modal is now a variant of the component
rather than the default.