Shorten the environment docker tag input string
This improves the UX for entering the docker tag by allowing more of
the relevant bits of the string to fit into the text field's viewport
without munging the layouts.

You can still enter "whole" docker tags, but by default it prepends
the "materialize/environmentd:" prefix if the : is missing.