Prepare for v0.27.0 release
Tomorrow at 4pm we'll roll out environmentd v0.27.0 to Materialize
Cloud. This new `environmentd` commit gets us ready for that release by
upgrading to an unstable SHA that includes all the breaking changes that
will be present in that release.

Once this change rolls out to production, we'll need to delete
everyone's environment, since this commit doesn't include migration code
for the backwards incompatible changes.

If all goes according to plan, any environments that get recreated at
this commit will be smoothly upgradable to v0.27.0.

This commit also removes the ability to set the environmentd version
from the UI. This is to set the right mentality: releases are controlled
by the cloud team, not the end user. The ability to override the
environmentd version via the API remains, though, and I'll add a section
to the playbook for our developers about how to use curl to add a
version override for an environment as necessary.