Remove the ability to destroy a region
When we ship to customers, enabling a region will be a one-way street.
There is too much important state associated with a region (e.g.,
billing) to allow customers to destroy them.

This commit removes just the web UI for deleting regions; not the API
endpoint. THe API endpoint should go away eventually too, I think, but
doing so will break the E2E platform test until we come up with another
way to delete regions.

The idea is to build the habit of nursing regions back to health when
they fail, just like we'll do for our real customers.

Fixes #3361 by removing the feature.