Autogenerate new app password for users if they don't have one already
If they do have one, the box is still there, it just doesn't show a key
(but the manage keys (link to app passwords page) and create new buttons
are still there)

This is naively checking if there are _any_ keys already there. It's not
checking if they've _ever_ had a key, and there's no distinguishing auto-
generated keys from other ones. Default name is My MZ key (with a number
appended if it's not the first key).

Create new button is disabled if we're currently displaying a key. Why
enable unnecessary key spam (we will have plenty enough as it is)

Remove memo from app password page that wasn't doing anything useful

Ditch the additional steps box (since a new tutorial is in the works)

Return empty string if we didn't get a clientId and secret back from frontegg despite there being a successDialog