region-controller
Creates a new region-controller deployment in the system cluster of each
AWS region, which is responsible for reconciling
`EnvironmentAssignments` and creating `Environment` resources in the
environment clusters.

Responsibility for creation and destruction of regional state is moved
from the environment-controller to the region-controller.
This includes cockroach databases/schemas and S3 data.

The database name is now based on the environment name, rather than its
UID, so that it may be stable when moved across environment clusters.

Most of the kuttl tests have been moved around to avoid race conditions
and to fix their cockroach auth.

Also fixed a few bugs in bin/serve that previously made it totally
non-functional.

fix lint and cargo tests

pass e2e tests