Make new region enable/select flow
https://github.com/MaterializeInc/cloud/issues/3604

Now _all_ possible regions appear in the dropdown. If the
user has enabled at least one region, it will appears the default
"current" region. If you select a non-enabled region in the dropdown,
you get a screen with a big ol' button to activate it.

Refactor how we get all and active regions from the APIs. Includes
helper methods to go from a region -> assignments and an assignment ->
environments, which are composed in useAvailableEnvironments and can
be called by themselves by other components as needed.

Someday when we have caching we'll probably have a central store of
all regions, environments, and their statuses, and this hook will
populate it with polling. But not yet.