Remove platform_enabled flag from Django
The platform_enabled flag in Django has largely been supplanted by the
Frontegg-managed "MaterializePlatform" and "MaterializePlatformAdmin"
roles. Port over the last few usages of the `platform_enabled` flag to
check for the existence of a `MaterializePlatform*` role instead, so
that we don't have two separate flags in two separate systems
determining platform access.