Show env status in dot in region select dropdown
Introduces react-select as a dependency (which allows us to style dropdown options).

Note: I observe some jankiness with the new library. If you're hovering over an option
and the refetcher refetches the data, even if there are no changes that ought to affect
the component (the region, provider, and color are identical) the options will 'flicker'
and you'll lose your hover state. Have not yet determined if this is something we can
fix (I added a `key` to the option objects but that doesn't appear to have had any
effect).

I don't see any real alternative to this library though--not much out there that's well
maintained, at least that I've found so far.