Re-do platform tests to be parallel per region & actually work
The platform test didn't test anything previously (see
https://github.com/MaterializeInc/cloud/issues/3974) - they assumed
that there was going to be a "platform table" which isn't the case in
all cases, and even if there was, the test skipped them.

So, instead do the following:
* Make one test per region, which will be run one per worker (so, in
  parallel!)
* Have each test determine the state that applies for its region via
  racing playwright promises (so that we don't have to use flakey
  timeouts) and take action to bring the region online.
* When that has happened, test the region & clean up afterward.

The test doesn't deactivate the region again, which would be really
nice to have but isn't supported at the moment. Sooo I guess we'll end
up testing stale regions a lot for now.