Stop cycling through incorrect UI states while getting data
The "we're not ready yet" instructions no longer flash on enabled envs.

The "region enabled" notification after pressing the button to activate a new
region should no longer appear before the home screen updates to
loading/starting/enabled UI.

When it's loading but we really don't know the status yet, it shows a spinner.
But that shouldn't be there long.