Rip out unnecessary code instead of adding unnecessary perms
The tenant token write/read permissions are only needed because we
have a switch in the app passwords page for stuff that we don't
support yet, so let's rip that code out and leave the permissions gone
for now, instead.

Due to a bug in Frontegg we have to leave the tenant API read
permission.