Remove Django organization API
Part of the Djang-no epic (#2872).

This commit removes the public Django API for retrieving organization
metadata. The only purpose this served was displaying a trial countdown
bubble in the nav bar; we're willing to live without that for now.

The deployment limit and enable dangerous features fields are also
removed from the Organization model, as they were only relevant for the
legacy cloud and are no longer referenced by any code.

Fix #2874.