Remove CSP when running frontend locally
Configuring CSP when running the frontend locally entails quite a bit of
complexity and doesn't seem to buy us anything. I think, in theory, it
could be useful for debugging the production CSP configuration, but in
practice it is different enough that you need to do CSP debugging on a
personal stack anyway.