Rewrite useAvailableEnvironments hook
This commit substantiallly remodels how region and environment state is
stored in Recoil and fetched from the region controllers and environment
controllers. The end result is more straightforward code that, in my
testing, does not exhibit the occasional flickers or stuck loading
screens that plagued the old implementation.

The basic idea is to avoid denormalizing state. There is one source of
truth for the state of an environment: the ENVIRONMENTS atom family,
which stores, for each environment, a discriminated union representing
the environment's state. Components that wish to access only loaded
environments' states, or just the current environment's state, do so via
Recoil selectors, which automatically update when the base atoms change.

Having just one atom family to update drastically simplifies the
useAvailableEnvironments hook, which previously had to juggle
simultaneous updates to multiple atoms storing different subsets of the
data.

Fix #4397.