De-cardify connect view on home page
Now instead of everything being inside a card, it's flat against the page
background like the empty-noun pages.

Reorders the connection info and app passwords.

Still uses old components (like card tabs) rather than e.g. making a
tabbed version of CodeBlock, so it's not a faithful rendition of the
current state of the designs. But it's a step towards them. Certainly
puts things more in line with the empty-noun pages' layout!