Update LaunchDarkly keys
I originally re-used some environments that had already been created,
but it was confusing, since environments have a key that can't be
changed which didn't match the name of the environment, so we decided to
create new ones.

I also decided personal stacks should be using the development
environment, not staging.

Discussion that led to this change:
https://www.notion.so/Using-LaunchDarkly-45cf26682e1b4d1d87325d04f5885725?d=7323f63f627148d3b1665e4ecb47f67e#6f29385f73a444a798da5c882b2f1e1d