Extend executeSql to allow multiple queries in one request
Also, move the decision about whether to pass
`SET cluster=mz_introspection` up to the caller. Update
`useSql` (which wraps `executeSql`, turning into a react hook) to
allow it to handle multiple queries as well. Rename it to
`useSqlMany`, and introduce a new `useSql` for compatibility that
retains the old behavior.