frontend: treat replica IDs as strings
To prepare for an upcoming change in the database that makes replica IDs
in the system tables represented as (GlobalId) strings rather than
integers, this commit adjusts the frontend code to tread replica IDs as
strings. This strategy should work with both the current and the
upcoming replica ID format, so no further migration is necessary.