Create view for list of connections
Creates a page that shows a table of connections and its data fetched from useConnections. People are able to filter by name, database, and schema. Also adds a tooltip to see what schema/database a connection is a part of. Handles the state where there was an error, it's loading, there are no connections, and there are no connections after searching.