Enforce per-organization region limit using LaunchDarkly
Use a new feature flag called "max_environments" which limits the number
of environments ("regions") that a user can enable. When an org  is at
this limit, the "enable region" button will be disabled for all users,
with title text explaining the situation. The UX of this leaves
something to be desired, but getting the control in place is the
priority.

The limit is only enforced client side, so users can circumvent this
restriction via direct API calls or the `mz` CLI, but that's okay: we
built a Metabase dashboard (see MaterializeInc/analytics#270) that will
catch these users and allow us to manually intervene.

Fix #28.
Fix #285.