Include group ID in Segment events
Segment "groups" are what we call "organizations". This commit sends the
user's currently active organization along with every `track` and `page`
call, so that our analytics can determine to which organization each
event belongs. To avoid code duplication, this commit introduces a new
`useSegment` hook which returns a wrapped Segment client that
automatically attaches the organization ID.

This commit also merges the AnalyticsOnEveryPage into the main router
component to match how the LaunchDarkly client is configured.