Improved useSql hook family loading states
Currently we rely on null vs empty array to distinguish between empty
and not yet loaded states, but this kind of implicit state leads to
bugs. Instead, we can track that we have actually attempted to load the
data at least once, and the empty data type can be empty array, rather
than null.

Both loading and isInitiallyLoading start out as true for useSql and
useSqlMany after this change.